New structure: GraphicHelpers and Invoices.Graphics
