+++ lib/siwapp_web/templates/page/index.html.heex
@@ -1,1 +1,1 @@
-<%= Graphics.amount_per_day_chart() %>
+<%= Invoices.Graphics.amount_per_day_chart() %>
