+++ lib/siwapp_web/views/page_view.ex
@@ -1,5 +1,5 @@
 defmodule SiwappWeb.PageView do
   use SiwappWeb, :view
 
-  alias SiwappWeb.GraphicHelpers, as: Graphics
+  alias Siwapp.Invoices
 end
